requirements.txt
dj-database-url==0.5.0
dj-static==0.0.6
Django==1.8.19
django-webpack-loader==0.2.4
djangorestframework==3.3.2
gunicorn==19.9.0
static3==0.7.0
psycopg2>=2.6.1